requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
matplotlib
scikit-image
scipy
pillow